changes to tests for using cdi 2.7
