+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.0] - 
+### Changed
+- Extracted email verification into its own recipe
+
 ## [3.2.0] - 2021-01-26
 ### Changed
 - Normalises email by making it all lower case
