+++ build.gradle
@@ -10,7 +10,7 @@ plugins {
     id 'application'
 }
 
-version = "3.2.0"
+version = "4.0.0"
 
 repositories {
     jcenter()
