+++ src/test/java/io/supertokens/test/GeneratePasswordResetTokenAPITest2_7.java
@@ -18,7 +18,6 @@ package io.supertokens.test;
 
 import com.google.gson.JsonObject;
 import io.supertokens.ProcessState;
-import io.supertokens.emailpassword.EmailPassword;
 import io.supertokens.pluginInterface.STORAGE_TYPE;
 import io.supertokens.storageLayer.StorageLayer;
 import org.junit.AfterClass;
@@ -29,7 +28,7 @@ import org.junit.rules.TestRule;
 
 import static org.junit.Assert.*;
 
-public class GenerateEmailVerificationTokenAPITest2_6 {
+public class GeneratePasswordResetTokenAPITest2_7 {
 
     @Rule
     public TestRule watchman = Utils.getOnFailure();
@@ -44,6 +43,7 @@ public class GenerateEmailVerificationTokenAPITest2_6 {
         Utils.reset();
     }
 
+    //Check for bad input (missing fields)
     @Test
     public void testBadInput() throws Exception {
         String[] args = {"../"};
@@ -55,14 +55,13 @@ public class GenerateEmailVerificationTokenAPITest2_6 {
             return;
         }
 
-        // not passing userID
         {
             try {
                 io.supertokens.test.httpRequest.HttpRequest
                         .sendJsonPOSTRequest(process.getProcess(), "",
-                                "http://localhost:3567/recipe/user/email/verify/token", null, 1000,
+                                "http://localhost:3567/recipe/user/password/reset/token", null, 1000,
                                 1000,
-                                null, Utils.getCdiVersion2_6ForTests());
+                                null, Utils.getCdiVersion2_7ForTests());
                 throw new Exception("Should not come here");
             } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
                 assertTrue(e.statusCode == 400 &&
@@ -75,9 +74,28 @@ public class GenerateEmailVerificationTokenAPITest2_6 {
             try {
                 io.supertokens.test.httpRequest.HttpRequest
                         .sendJsonPOSTRequest(process.getProcess(), "",
-                                "http://localhost:3567/recipe/user/email/verify/token", requestBody, 1000,
+                                "http://localhost:3567/recipe/user/password/reset/token", requestBody, 1000,
                                 1000,
-                                null, Utils.getCdiVersion2_6ForTests());
+                                null, Utils.getCdiVersion2_7ForTests());
+                throw new Exception("Should not come here");
+            } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
+                assertTrue(e.statusCode == 400 &&
+                        e.getMessage()
+                                .equals("Http error. Status Code: 400. Message: Field name 'userId' is invalid in " +
+                                        "JSON input"));
+
+            }
+        }
+
+        {
+            JsonObject requestBody = new JsonObject();
+            requestBody.add("userId", null);
+            try {
+                io.supertokens.test.httpRequest.HttpRequest
+                        .sendJsonPOSTRequest(process.getProcess(), "",
+                                "http://localhost:3567/recipe/user/password/reset/token", requestBody, 1000,
+                                1000,
+                                null, Utils.getCdiVersion2_7ForTests());
                 throw new Exception("Should not come here");
             } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
                 assertTrue(e.statusCode == 400 &&
@@ -87,23 +105,6 @@ public class GenerateEmailVerificationTokenAPITest2_6 {
             }
         }
 
-        {
-            JsonObject requestBody = new JsonObject();
-            requestBody.addProperty("userId", 12345);
-            try {
-                io.supertokens.test.httpRequest.HttpRequest
-                        .sendJsonPOSTRequest(process.getProcess(), "",
-                                "http://localhost:3567/recipe/user/email/verify/token", requestBody, 1000,
-                                1000,
-                                null, Utils.getCdiVersion2_6ForTests());
-                throw new Exception("Should not come here");
-            } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
-                assertTrue(e.statusCode == 400 &&
-                        e.getMessage()
-                                .equals("Http error. Status Code: 400. Message: Field name 'userId' is invalid in " +
-                                        "JSON input"));
-            }
-        }
         process.kill();
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
     }
@@ -120,32 +121,32 @@ public class GenerateEmailVerificationTokenAPITest2_6 {
             return;
         }
 
-        JsonObject signUpResponse = Utils.signUpRequest_2_5(process, "random@gmail.com", "validPass123");
+        JsonObject signUpResponse = Utils.signUpRequest_2_4(process, "random@gmail.com", "validPass123");
         assertEquals(signUpResponse.get("status").getAsString(), "OK");
         assertEquals(signUpResponse.entrySet().size(), 2);
+        String userId = signUpResponse.getAsJsonObject("user").get("id").getAsString();
 
-        String userId = signUpResponse.get("user").getAsJsonObject().get("id").getAsString();
         JsonObject requestBody = new JsonObject();
         requestBody.addProperty("userId", userId);
 
-
         JsonObject response = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "",
-                        "http://localhost:3567/recipe/user/email/verify/token", requestBody, 1000,
+                        "http://localhost:3567/recipe/user/password/reset/token", requestBody, 1000,
                         1000,
-                        null, Utils.getCdiVersion2_6ForTests());
+                        null, Utils.getCdiVersion2_7ForTests());
 
-        assertEquals(response.entrySet().size(), 2);
         assertEquals(response.get("status").getAsString(), "OK");
         assertNotNull(response.get("token"));
+        assertEquals(response.entrySet().size(), 2);
 
         process.kill();
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
     }
 
-    // Check for all types of output
+    //Check for all types of output
+    // Failure condition: passing a valid userId will cause the test to fail
     @Test
-    public void testAllTypesOfOutput() throws Exception {
+    public void testForAllTypesOfOutput() throws Exception {
         String[] args = {"../"};
 
         TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
@@ -155,49 +156,17 @@ public class GenerateEmailVerificationTokenAPITest2_6 {
             return;
         }
 
-        // user does not exist
-        {
-            JsonObject requestBody = new JsonObject();
-            requestBody.addProperty("userId", "randomUserId");
+        JsonObject requestBody = new JsonObject();
+        requestBody.addProperty("userId", "randomUserId");
 
-            JsonObject response = io.supertokens.test.httpRequest.HttpRequest
-                    .sendJsonPOSTRequest(process.getProcess(), "",
-                            "http://localhost:3567/recipe/user/email/verify/token", requestBody, 1000,
-                            1000,
-                            null, Utils.getCdiVersion2_6ForTests());
-            assertEquals(response.get("status").getAsString(), "UNKNOWN_USER_ID_ERROR");
-        }
+        JsonObject response = io.supertokens.test.httpRequest.HttpRequest
+                .sendJsonPOSTRequest(process.getProcess(), "",
+                        "http://localhost:3567/recipe/user/password/reset/token", requestBody, 1000,
+                        1000,
+                        null, Utils.getCdiVersion2_7ForTests());
 
-        // user exists but email already verified
-        {
-            JsonObject signUpResponse = Utils.signUpRequest_2_5(process, "random@gmail.com", "validPass123");
-            assertEquals(signUpResponse.get("status").getAsString(), "OK");
-            assertEquals(signUpResponse.entrySet().size(), 2);
-
-            String userId = signUpResponse.get("user").getAsJsonObject().get("id").getAsString();
-            JsonObject requestBody = new JsonObject();
-            requestBody.addProperty("userId", userId);
-
-
-            JsonObject response = io.supertokens.test.httpRequest.HttpRequest
-                    .sendJsonPOSTRequest(process.getProcess(), "",
-                            "http://localhost:3567/recipe/user/email/verify/token", requestBody, 1000,
-                            1000,
-                            null, Utils.getCdiVersion2_6ForTests());
-
-            assertEquals(response.entrySet().size(), 2);
-            assertEquals(response.get("status").getAsString(), "OK");
-            assertNotNull(response.get("token"));
-
-            EmailPassword.verifyEmail(process.getProcess(), response.get("token").getAsString());
-
-            JsonObject response2 = io.supertokens.test.httpRequest.HttpRequest
-                    .sendJsonPOSTRequest(process.getProcess(), "",
-                            "http://localhost:3567/recipe/user/email/verify/token", requestBody, 1000,
-                            1000,
-                            null, Utils.getCdiVersion2_6ForTests());
-            assertEquals(response2.get("status").getAsString(), "EMAIL_ALREADY_VERIFIED_ERROR");
-        }
+        assertEquals(response.get("status").getAsString(), "UNKNOWN_USER_ID_ERROR");
+        assertEquals(response.entrySet().size(), 1);
 
         process.kill();
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
