+++ src/test/java/io/supertokens/test/IsEmailVerifiedAPITest2_7.java
@@ -30,7 +30,7 @@ import java.util.HashMap;
 
 import static org.junit.Assert.*;
 
-public class IsEmailVerifiedAPITest2_6 {
+public class IsEmailVerifiedAPITest2_7 {
 
     @Rule
     public TestRule watchman = Utils.getOnFailure();
@@ -60,7 +60,7 @@ public class IsEmailVerifiedAPITest2_6 {
             try {
                 io.supertokens.test.httpRequest.HttpRequest
                         .sendGETRequest(process.getProcess(), "", "http://localhost:3567/recipe/user/email/verify",
-                                null, 1000, 1000, null, Utils.getCdiVersion2_6ForTests());
+                                null, 1000, 1000, null, Utils.getCdiVersion2_7ForTests());
                 throw new Exception("Should not come here");
             } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
                 assertTrue(e.statusCode == 400 &&
@@ -78,7 +78,7 @@ public class IsEmailVerifiedAPITest2_6 {
 
                 io.supertokens.test.httpRequest.HttpRequest
                         .sendGETRequest(process.getProcess(), "", "http://localhost:3567/recipe/user/email/verify",
-                                map, 1000, 1000, null, Utils.getCdiVersion2_6ForTests());
+                                map, 1000, 1000, null, Utils.getCdiVersion2_7ForTests());
                 throw new Exception("Should not come here");
             } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
                 assertTrue(e.statusCode == 400 &&
@@ -115,7 +115,7 @@ public class IsEmailVerifiedAPITest2_6 {
 
         JsonObject verifyResponse = io.supertokens.test.httpRequest.HttpRequest
                 .sendGETRequest(process.getProcess(), "", "http://localhost:3567/recipe/user/email/verify",
-                        map, 1000, 1000, null, Utils.getCdiVersion2_6ForTests());
+                        map, 1000, 1000, null, Utils.getCdiVersion2_7ForTests());
         assertEquals(verifyResponse.entrySet().size(), 2);
         assertEquals(verifyResponse.get("status").getAsString(), "OK");
         assertEquals(verifyResponse.get("isVerified").getAsBoolean(), false);
@@ -129,7 +129,7 @@ public class IsEmailVerifiedAPITest2_6 {
                 .sendJsonPOSTRequest(process.getProcess(), "",
                         "http://localhost:3567/recipe/user/email/verify/token", requestBody, 1000,
                         1000,
-                        null, Utils.getCdiVersion2_6ForTests());
+                        null, Utils.getCdiVersion2_7ForTests());
 
         assertEquals(response.entrySet().size(), 2);
         assertEquals(response.get("status").getAsString(), "OK");
@@ -143,7 +143,7 @@ public class IsEmailVerifiedAPITest2_6 {
                 .sendJsonPOSTRequest(process.getProcess(), "",
                         "http://localhost:3567/recipe/user/email/verify", verifyResponseBody, 1000,
                         1000,
-                        null, Utils.getCdiVersion2_6ForTests());
+                        null, Utils.getCdiVersion2_7ForTests());
 
         assertEquals(response2.entrySet().size(), 2);
         assertEquals(response2.get("status").getAsString(), "OK");
@@ -154,7 +154,7 @@ public class IsEmailVerifiedAPITest2_6 {
 
         verifyResponse = io.supertokens.test.httpRequest.HttpRequest
                 .sendGETRequest(process.getProcess(), "", "http://localhost:3567/recipe/user/email/verify",
-                        map, 1000, 1000, null, Utils.getCdiVersion2_6ForTests());
+                        map, 1000, 1000, null, Utils.getCdiVersion2_7ForTests());
         assertEquals(verifyResponse.entrySet().size(), 2);
         assertEquals(verifyResponse.get("status").getAsString(), "OK");
         assertEquals(verifyResponse.get("isVerified").getAsBoolean(), true);
@@ -183,7 +183,7 @@ public class IsEmailVerifiedAPITest2_6 {
                     .sendGETRequest(process.getProcess(), "",
                             "http://localhost:3567/recipe/user/email/verify", map, 1000,
                             1000,
-                            null, Utils.getCdiVersion2_6ForTests());
+                            null, Utils.getCdiVersion2_7ForTests());
             assertEquals(response2.get("status").getAsString(), "UNKNOWN_USER_ID_ERROR");
             assertEquals(response2.entrySet().size(), 1);
         }
