+++ src/test/java/io/supertokens/test/SessionUserAPITest2_7.java
@@ -31,7 +31,7 @@ import java.util.HashMap;
 import static org.junit.Assert.*;
 
 
-public class SessionUserAPITest2_6 {
+public class SessionUserAPITest2_7 {
     @Rule
     public TestRule watchman = Utils.getOnFailure();
 
@@ -58,7 +58,7 @@ public class SessionUserAPITest2_6 {
             io.supertokens.test.httpRequest.HttpRequest
                     .sendGETRequest(process.getProcess(), "", "http://localhost:3567/recipe/session/user", null, 1000,
                             1000,
-                            null, Utils.getCdiVersion2_6ForTests());
+                            null, Utils.getCdiVersion2_7ForTests());
             fail();
         } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
             assertTrue(e.statusCode == 400 && e.getMessage()
@@ -73,7 +73,7 @@ public class SessionUserAPITest2_6 {
             io.supertokens.test.httpRequest.HttpRequest
                     .sendGETRequest(process.getProcess(), "", "http://localhost:3567/recipe/session/user", map, 1000,
                             1000,
-                            null, Utils.getCdiVersion2_6ForTests());
+                            null, Utils.getCdiVersion2_7ForTests());
             fail();
         } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
             assertTrue(e.statusCode == 400 && e.getMessage()
@@ -112,7 +112,7 @@ public class SessionUserAPITest2_6 {
         JsonObject response = io.supertokens.test.httpRequest.HttpRequest
                 .sendGETRequest(process.getProcess(), "", "http://localhost:3567/recipe/session/user", map, 1000, 1000,
                         null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
 
         assertEquals(response.get("status").getAsString(), "OK");
         assertEquals(0, response.get("sessionHandles").getAsJsonArray().size());
@@ -121,7 +121,7 @@ public class SessionUserAPITest2_6 {
         JsonObject sessionCreatedResponse = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session",
                         new JsonParser().parse(createSessionJsonInput).getAsJsonObject(), 1000, 1000, null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
 
         assertEquals("OK", sessionCreatedResponse.get("status").getAsString());
 
@@ -130,7 +130,7 @@ public class SessionUserAPITest2_6 {
         response = io.supertokens.test.httpRequest.HttpRequest
                 .sendGETRequest(process.getProcess(), "", "http://localhost:3567/recipe/session/user", map, 1000, 1000,
                         null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
 
         assertEquals(response.get("status").getAsString(), "OK");
         assertEquals(response.entrySet().size(), 2);
@@ -144,7 +144,7 @@ public class SessionUserAPITest2_6 {
         JsonObject sessionCreatedResponse1 = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session",
                         new JsonParser().parse(createSessionJsonInput).getAsJsonObject(), 1000, 1000, null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
 
         assertEquals(sessionCreatedResponse1.get("status").getAsString(), "OK");
 
@@ -152,14 +152,14 @@ public class SessionUserAPITest2_6 {
         JsonObject sessionCreatedResponse2 = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session",
                         new JsonParser().parse(createSessionJsonInput).getAsJsonObject(), 1000, 1000, null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
 
         assertEquals(sessionCreatedResponse2.get("status").getAsString(), "OK");
 
         JsonObject multiResponse = io.supertokens.test.httpRequest.HttpRequest
                 .sendGETRequest(process.getProcess(), "", "http://localhost:3567/recipe/session/user", map, 1000, 1000,
                         null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
 
         assertEquals(multiResponse.get("status").getAsString(), "OK");
         assertEquals(multiResponse.entrySet().size(), 2);
@@ -228,7 +228,7 @@ public class SessionUserAPITest2_6 {
         JsonObject user1Response1 = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session",
                         new JsonParser().parse(userJsonInput1).getAsJsonObject(), 1000, 1000, null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
 
         assertEquals("OK", user1Response1.get("status").getAsString());
 
@@ -236,7 +236,7 @@ public class SessionUserAPITest2_6 {
         JsonObject user1Response2 = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session",
                         new JsonParser().parse(userJsonInput1).getAsJsonObject(), 1000, 1000, null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
 
         assertEquals("OK", user1Response2.get("status").getAsString());
 
@@ -244,7 +244,7 @@ public class SessionUserAPITest2_6 {
         JsonObject user1Response3 = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session",
                         new JsonParser().parse(userJsonInput1).getAsJsonObject(), 1000, 1000, null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
 
         assertEquals("OK", user1Response3.get("status").getAsString());
 
@@ -256,7 +256,7 @@ public class SessionUserAPITest2_6 {
         JsonObject user2Response1 = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session",
                         new JsonParser().parse(userJsonInput2).getAsJsonObject(), 1000, 1000, null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
 
         assertEquals("OK", user2Response1.get("status").getAsString());
 
@@ -265,7 +265,7 @@ public class SessionUserAPITest2_6 {
         JsonObject user2Response2 = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session",
                         new JsonParser().parse(userJsonInput2).getAsJsonObject(), 1000, 1000, null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
 
         assertEquals("OK", user2Response2.get("status").getAsString());
 
@@ -274,7 +274,7 @@ public class SessionUserAPITest2_6 {
         JsonObject user2Response3 = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session",
                         new JsonParser().parse(userJsonInput2).getAsJsonObject(), 1000, 1000, null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
 
         assertEquals("OK", user2Response3.get("status").getAsString());
 
@@ -286,7 +286,7 @@ public class SessionUserAPITest2_6 {
         JsonObject user3Response1 = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session",
                         new JsonParser().parse(userJsonInput3).getAsJsonObject(), 1000, 1000, null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
 
         assertEquals("OK", user3Response1.get("status").getAsString());
 
@@ -295,7 +295,7 @@ public class SessionUserAPITest2_6 {
         JsonObject user3Response2 = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session",
                         new JsonParser().parse(userJsonInput3).getAsJsonObject(), 1000, 1000, null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
 
         assertEquals("OK", user3Response2.get("status").getAsString());
 
@@ -304,7 +304,7 @@ public class SessionUserAPITest2_6 {
         JsonObject user3Response3 = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session",
                         new JsonParser().parse(userJsonInput3).getAsJsonObject(), 1000, 1000, null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
 
         assertEquals("OK", user3Response3.get("status").getAsString());
 
@@ -316,7 +316,7 @@ public class SessionUserAPITest2_6 {
         JsonObject multiResponse = io.supertokens.test.httpRequest.HttpRequest
                 .sendGETRequest(process.getProcess(), "", "http://localhost:3567/recipe/session/user", map, 1000, 1000,
                         null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
 
         assertEquals("OK", multiResponse.get("status").getAsString());
         assertEquals(multiResponse.get("sessionHandles").getAsJsonArray().size(), 3);
@@ -335,7 +335,7 @@ public class SessionUserAPITest2_6 {
         multiResponse = io.supertokens.test.httpRequest.HttpRequest
                 .sendGETRequest(process.getProcess(), "", "http://localhost:3567/recipe/session/user", map, 1000, 1000,
                         null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
         assertEquals("OK", multiResponse.get("status").getAsString());
 
         sessionArray = multiResponse.get("sessionHandles").getAsJsonArray();
@@ -351,7 +351,7 @@ public class SessionUserAPITest2_6 {
         multiResponse = io.supertokens.test.httpRequest.HttpRequest
                 .sendGETRequest(process.getProcess(), "", "http://localhost:3567/recipe/session/user", map, 1000, 1000,
                         null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
         assertEquals("OK", multiResponse.get("status").getAsString());
 
         sessionArray = multiResponse.get("sessionHandles").getAsJsonArray();
