+++ src/test/java/io/supertokens/test/SignInAPITest2_7.java
@@ -19,7 +19,6 @@ package io.supertokens.test;
 import com.google.gson.JsonObject;
 import io.supertokens.ProcessState;
 import io.supertokens.pluginInterface.STORAGE_TYPE;
-import io.supertokens.pluginInterface.emailpassword.UserInfo;
 import io.supertokens.storageLayer.StorageLayer;
 import org.junit.AfterClass;
 import org.junit.Before;
@@ -32,13 +31,14 @@ import static org.junit.Assert.*;
 /*
  * TODO:
  *  - Check for bad input (missing fields)
- *  - Check good input works and that user is there in db (and then call sign in)
- *  - Test the normalise email function
- *  - Test that only the normalised email is saved in the db
- *  - Test that giving an empty password throws a bad input error
+ *  - Check good input works
+ *  - Test that sign in with unnormalised email like Test@gmail.com should also work
+ *  - Test that giving an empty password, empty email, invalid email, random email or wrong password throws a wrong
+ *      credentials error
+ *  - Test that an empty password yields a WRONG_CREDENTIALS_ERROR output.
  * */
 
-public class SignUpAPITest2_6 {
+public class SignInAPITest2_7 {
 
     @Rule
     public TestRule watchman = Utils.getOnFailure();
@@ -69,9 +69,9 @@ public class SignUpAPITest2_6 {
             try {
                 io.supertokens.test.httpRequest.HttpRequest
                         .sendJsonPOSTRequest(process.getProcess(), "",
-                                "http://localhost:3567/recipe/signup", null, 1000,
+                                "http://localhost:3567/recipe/signin", null, 1000,
                                 1000,
-                                null, Utils.getCdiVersion2_6ForTests());
+                                null, Utils.getCdiVersion2_7ForTests());
                 throw new Exception("Should not come here");
             } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
                 assertTrue(e.statusCode == 400 &&
@@ -85,9 +85,9 @@ public class SignUpAPITest2_6 {
             try {
                 io.supertokens.test.httpRequest.HttpRequest
                         .sendJsonPOSTRequest(process.getProcess(), "",
-                                "http://localhost:3567/recipe/signup", requestBody, 1000,
+                                "http://localhost:3567/recipe/signin", requestBody, 1000,
                                 1000,
-                                null, Utils.getCdiVersion2_6ForTests());
+                                null, Utils.getCdiVersion2_7ForTests());
                 throw new Exception("Should not come here");
             } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
                 assertTrue(e.statusCode == 400 &&
@@ -102,9 +102,9 @@ public class SignUpAPITest2_6 {
             try {
                 io.supertokens.test.httpRequest.HttpRequest
                         .sendJsonPOSTRequest(process.getProcess(), "",
-                                "http://localhost:3567/recipe/signup", requestBody, 1000,
+                                "http://localhost:3567/recipe/signin", requestBody, 1000,
                                 1000,
-                                null, Utils.getCdiVersion2_6ForTests());
+                                null, Utils.getCdiVersion2_7ForTests());
                 throw new Exception("Should not come here");
             } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
                 assertTrue(e.statusCode == 400 &&
@@ -118,7 +118,7 @@ public class SignUpAPITest2_6 {
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
     }
 
-    //Check good input works and that user is there in db (and then call sign in)
+    //Check good input works
     @Test
     public void testGoodInput() throws Exception {
         String[] args = {"../"};
@@ -134,14 +134,7 @@ public class SignUpAPITest2_6 {
         assertEquals(signUpResponse.get("status").getAsString(), "OK");
         assertEquals(signUpResponse.entrySet().size(), 2);
 
-        JsonObject signUpUser = signUpResponse.get("user").getAsJsonObject();
-        assertEquals(signUpUser.get("email").getAsString(), "random@gmail.com");
-        assertNotNull(signUpUser.get("id"));
-
-        UserInfo user = StorageLayer.getEmailPasswordStorage(process.getProcess())
-                .getUserInfoUsingEmail("random@gmail.com");
-        assertEquals(user.email, signUpUser.get("email").getAsString());
-        assertEquals(user.id, signUpUser.get("id").getAsString());
+        JsonObject userInfo = signUpResponse.get("user").getAsJsonObject();
 
         JsonObject responseBody = new JsonObject();
         responseBody.addProperty("email", "random@gmail.com");
@@ -151,15 +144,15 @@ public class SignUpAPITest2_6 {
                 .sendJsonPOSTRequest(process.getProcess(), "",
                         "http://localhost:3567/recipe/signin", responseBody, 1000,
                         1000,
-                        null, Utils.getCdiVersion2_6ForTests());
+                        null, Utils.getCdiVersion2_7ForTests());
 
         assertEquals(signInResponse.get("status").getAsString(), "OK");
         assertEquals(signInResponse.entrySet().size(), 2);
 
         assertEquals(signInResponse.get("user").getAsJsonObject().get("id").getAsString(),
-                signUpUser.get("id").getAsString());
+                userInfo.get("id").getAsString());
         assertEquals(signInResponse.get("user").getAsJsonObject().get("email").getAsString(),
-                signUpUser.get("email").getAsString());
+                userInfo.get("email").getAsString());
         signInResponse.get("user").getAsJsonObject().get("timeJoined").getAsLong();
         assertEquals(signInResponse.get("user").getAsJsonObject().entrySet().size(), 3);
 
@@ -167,11 +160,9 @@ public class SignUpAPITest2_6 {
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
     }
 
-    // Test the normalise email function
-    // Test that only the normalised email is saved in the db
-    //Failure condition: If the email retrieved from the data is not normalised the test will fail
+    //Test that sign in with unnormalised email like Test@gmail.com should also work
     @Test
-    public void testTheNormaliseEmailFunction() throws Exception {
+    public void testThatUnnormalisedEmailShouldAlsoWork() throws Exception {
         String[] args = {"../"};
 
         TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
@@ -181,26 +172,39 @@ public class SignUpAPITest2_6 {
             return;
         }
 
-        JsonObject signUpResponse = Utils.signUpRequest_2_5(process, "RaNdOm@gmail.com", "validPass123");
+        JsonObject signUpResponse = Utils.signUpRequest_2_5(process, "test@gmail.com", "validPass123");
         assertEquals(signUpResponse.get("status").getAsString(), "OK");
         assertEquals(signUpResponse.entrySet().size(), 2);
 
-        JsonObject signUpUser = signUpResponse.get("user").getAsJsonObject();
-        assertEquals(signUpUser.get("email").getAsString(), "random@gmail.com");
-        assertNotNull(signUpUser.get("id"));
+        JsonObject userInfo = signUpResponse.get("user").getAsJsonObject();
 
-        UserInfo userInfo = StorageLayer.getEmailPasswordStorage(process.getProcess())
-                .getUserInfoUsingId(signUpUser.get("id").getAsString());
+        JsonObject responseBody = new JsonObject();
+        responseBody.addProperty("email", "Test@gmail.com");
+        responseBody.addProperty("password", "validPass123");
 
-        assertEquals(userInfo.email, "random@gmail.com");
+        JsonObject signInResponse = io.supertokens.test.httpRequest.HttpRequest
+                .sendJsonPOSTRequest(process.getProcess(), "",
+                        "http://localhost:3567/recipe/signin", responseBody, 1000,
+                        1000,
+                        null, Utils.getCdiVersion2_7ForTests());
+
+        assertEquals(signInResponse.get("status").getAsString(), "OK");
+        assertEquals(signInResponse.entrySet().size(), 2);
+        assertEquals(signInResponse.get("user").getAsJsonObject().get("id").getAsString(),
+                userInfo.get("id").getAsString());
+        assertEquals(signInResponse.get("user").getAsJsonObject().get("email").getAsString(),
+                userInfo.get("email").getAsString());
+        signInResponse.get("user").getAsJsonObject().get("timeJoined").getAsLong();
+        assertEquals(signInResponse.get("user").getAsJsonObject().entrySet().size(), 3);
 
         process.kill();
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
     }
 
-    //Test that giving an empty password throws a bad input error
+    //Test that giving an empty password, empty email, invalid email, random email or wrong password throws a wrong
+    // *      credentials error
     @Test
-    public void testEmptyPasswordThrowsBadInputError() throws Exception {
+    public void testInputsToSignInAPI() throws Exception {
         String[] args = {"../"};
 
         TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
@@ -210,42 +214,72 @@ public class SignUpAPITest2_6 {
             return;
         }
 
-        try {
-            Utils.signUpRequest_2_5(process, "random@gmail.com", "");
-            throw new Exception("Should not come here");
-        } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
-            assertTrue(e.statusCode == 400 &&
-                    e.getMessage().equals("Http error. Status Code: 400. Message: Password cannot be an empty string"));
+        {
+            JsonObject responseBody = new JsonObject();
+            responseBody.addProperty("email", "random@gmail.com");
+            responseBody.addProperty("password", "");
+
+            JsonObject signInResponse = io.supertokens.test.httpRequest.HttpRequest
+                    .sendJsonPOSTRequest(process.getProcess(), "",
+                            "http://localhost:3567/recipe/signin", responseBody, 1000,
+                            1000,
+                            null, Utils.getCdiVersion2_7ForTests());
+
+            assertEquals(signInResponse.get("status").getAsString(), "WRONG_CREDENTIALS_ERROR");
+            assertEquals(signInResponse.entrySet().size(), 1);
+        }
+
+        {
+            JsonObject responseBody = new JsonObject();
+            responseBody.addProperty("email", "");
+            responseBody.addProperty("password", "validPass123");
+
+            JsonObject signInResponse = io.supertokens.test.httpRequest.HttpRequest
+                    .sendJsonPOSTRequest(process.getProcess(), "",
+                            "http://localhost:3567/recipe/signin", responseBody, 1000,
+                            1000,
+                            null, Utils.getCdiVersion2_7ForTests());
+
+            assertEquals(signInResponse.get("status").getAsString(), "WRONG_CREDENTIALS_ERROR");
+            assertEquals(signInResponse.entrySet().size(), 1);
+        }
+
+        {
+            JsonObject responseBody = new JsonObject();
+            responseBody.addProperty("email", "random@gmail.com");
+            responseBody.addProperty("password", "randomPassword123");
+
+            JsonObject signInResponse = io.supertokens.test.httpRequest.HttpRequest
+                    .sendJsonPOSTRequest(process.getProcess(), "",
+                            "http://localhost:3567/recipe/signin", responseBody, 1000,
+                            1000,
+                            null, Utils.getCdiVersion2_7ForTests());
+
+            assertEquals(signInResponse.get("status").getAsString(), "WRONG_CREDENTIALS_ERROR");
+            assertEquals(signInResponse.entrySet().size(), 1);
+        }
+
+        {
+            JsonObject signUpResponse = Utils.signUpRequest_2_5(process, "test@gmail.com", "validPass123");
+            assertEquals(signUpResponse.get("status").getAsString(), "OK");
+            assertEquals(signUpResponse.entrySet().size(), 2);
+
+            JsonObject responseBody = new JsonObject();
+            responseBody.addProperty("email", "test@gmail.com");
+            responseBody.addProperty("password", "wrongPassword");
+
+            JsonObject signInResponse = io.supertokens.test.httpRequest.HttpRequest
+                    .sendJsonPOSTRequest(process.getProcess(), "",
+                            "http://localhost:3567/recipe/signin", responseBody, 1000,
+                            1000,
+                            null, Utils.getCdiVersion2_7ForTests());
+
+            assertEquals(signInResponse.get("status").getAsString(), "WRONG_CREDENTIALS_ERROR");
+            assertEquals(signInResponse.entrySet().size(), 1);
         }
 
         process.kill();
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
     }
 
-    @Test
-    public void testSignUpWithDupicateEMail() throws Exception {
-        String[] args = {"../"};
-
-        TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
-        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
-
-        if (StorageLayer.getStorage(process.getProcess()).getType() != STORAGE_TYPE.SQL) {
-            return;
-        }
-
-        JsonObject signUpResponse = Utils.signUpRequest_2_5(process, "random@gmail.com", "validPass123");
-        assertEquals(signUpResponse.get("status").getAsString(), "OK");
-        assertEquals(signUpResponse.entrySet().size(), 2);
-
-        JsonObject signUpUser = signUpResponse.get("user").getAsJsonObject();
-        assertEquals(signUpUser.get("email").getAsString(), "random@gmail.com");
-        assertNotNull(signUpUser.get("id"));
-
-        JsonObject response = Utils.signUpRequest_2_5(process, "random@gmail.com", "validPass123");
-        assertEquals(response.get("status").getAsString(), "EMAIL_ALREADY_EXISTS_ERROR");
-        assertEquals(response.entrySet().size(), 1);
-
-        process.kill();
-        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
-    }
 }
