+++ src/test/java/io/supertokens/test/VerifyEmailAPITest2_7.java
@@ -28,7 +28,7 @@ import org.junit.rules.TestRule;
 
 import static org.junit.Assert.*;
 
-public class GeneratePasswordResetTokenAPITest2_6 {
+public class VerifyEmailAPITest2_7 {
 
     @Rule
     public TestRule watchman = Utils.getOnFailure();
@@ -43,7 +43,6 @@ public class GeneratePasswordResetTokenAPITest2_6 {
         Utils.reset();
     }
 
-    //Check for bad input (missing fields)
     @Test
     public void testBadInput() throws Exception {
         String[] args = {"../"};
@@ -59,9 +58,9 @@ public class GeneratePasswordResetTokenAPITest2_6 {
             try {
                 io.supertokens.test.httpRequest.HttpRequest
                         .sendJsonPOSTRequest(process.getProcess(), "",
-                                "http://localhost:3567/recipe/user/password/reset/token", null, 1000,
+                                "http://localhost:3567/recipe/user/email/verify", null, 1000,
                                 1000,
-                                null, Utils.getCdiVersion2_6ForTests());
+                                null, Utils.getCdiVersion2_7ForTests());
                 throw new Exception("Should not come here");
             } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
                 assertTrue(e.statusCode == 400 &&
@@ -74,33 +73,33 @@ public class GeneratePasswordResetTokenAPITest2_6 {
             try {
                 io.supertokens.test.httpRequest.HttpRequest
                         .sendJsonPOSTRequest(process.getProcess(), "",
-                                "http://localhost:3567/recipe/user/password/reset/token", requestBody, 1000,
+                                "http://localhost:3567/recipe/user/email/verify", requestBody, 1000,
                                 1000,
-                                null, Utils.getCdiVersion2_6ForTests());
+                                null, Utils.getCdiVersion2_7ForTests());
                 throw new Exception("Should not come here");
             } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
                 assertTrue(e.statusCode == 400 &&
                         e.getMessage()
-                                .equals("Http error. Status Code: 400. Message: Field name 'userId' is invalid in " +
+                                .equals("Http error. Status Code: 400. Message: Field name 'method' is invalid in " +
                                         "JSON input"));
-
             }
         }
 
         {
             JsonObject requestBody = new JsonObject();
-            requestBody.add("userId", null);
+            requestBody.addProperty("method", "token");
+            requestBody.addProperty("token", 12345);
             try {
                 io.supertokens.test.httpRequest.HttpRequest
                         .sendJsonPOSTRequest(process.getProcess(), "",
-                                "http://localhost:3567/recipe/user/password/reset/token", requestBody, 1000,
+                                "http://localhost:3567/recipe/user/email/verify", requestBody, 1000,
                                 1000,
-                                null, Utils.getCdiVersion2_6ForTests());
+                                null, Utils.getCdiVersion2_7ForTests());
                 throw new Exception("Should not come here");
             } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
                 assertTrue(e.statusCode == 400 &&
                         e.getMessage()
-                                .equals("Http error. Status Code: 400. Message: Field name 'userId' is invalid in " +
+                                .equals("Http error. Status Code: 400. Message: Field name 'token' is invalid in " +
                                         "JSON input"));
             }
         }
@@ -121,32 +120,49 @@ public class GeneratePasswordResetTokenAPITest2_6 {
             return;
         }
 
-        JsonObject signUpResponse = Utils.signUpRequest_2_4(process, "random@gmail.com", "validPass123");
+        JsonObject signUpResponse = Utils.signUpRequest_2_5(process, "random@gmail.com", "validPass123");
         assertEquals(signUpResponse.get("status").getAsString(), "OK");
         assertEquals(signUpResponse.entrySet().size(), 2);
-        String userId = signUpResponse.getAsJsonObject("user").get("id").getAsString();
 
+        String userId = signUpResponse.get("user").getAsJsonObject().get("id").getAsString();
         JsonObject requestBody = new JsonObject();
         requestBody.addProperty("userId", userId);
 
+
         JsonObject response = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "",
-                        "http://localhost:3567/recipe/user/password/reset/token", requestBody, 1000,
+                        "http://localhost:3567/recipe/user/email/verify/token", requestBody, 1000,
                         1000,
-                        null, Utils.getCdiVersion2_6ForTests());
+                        null, Utils.getCdiVersion2_7ForTests());
 
+        assertEquals(response.entrySet().size(), 2);
         assertEquals(response.get("status").getAsString(), "OK");
         assertNotNull(response.get("token"));
-        assertEquals(response.entrySet().size(), 2);
+
+        JsonObject verifyResponseBody = new JsonObject();
+        verifyResponseBody.addProperty("method", "token");
+        verifyResponseBody.addProperty("token", response.get("token").getAsString());
+
+        JsonObject response2 = io.supertokens.test.httpRequest.HttpRequest
+                .sendJsonPOSTRequest(process.getProcess(), "",
+                        "http://localhost:3567/recipe/user/email/verify", verifyResponseBody, 1000,
+                        1000,
+                        null, Utils.getCdiVersion2_7ForTests());
+
+        assertEquals(response2.entrySet().size(), 2);
+        assertEquals(response2.get("status").getAsString(), "OK");
+
+        assertEquals(response2.get("user").getAsJsonObject().entrySet().size(), 3);
+        assertEquals(response2.get("user").getAsJsonObject().get("id").getAsString(), userId);
+        assertEquals(response2.get("user").getAsJsonObject().get("email").getAsString(), "random@gmail.com");
 
         process.kill();
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
     }
 
-    //Check for all types of output
-    // Failure condition: passing a valid userId will cause the test to fail
+    // Check for all types of output
     @Test
-    public void testForAllTypesOfOutput() throws Exception {
+    public void testAllTypesOfOutput() throws Exception {
         String[] args = {"../"};
 
         TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
@@ -156,19 +172,42 @@ public class GeneratePasswordResetTokenAPITest2_6 {
             return;
         }
 
-        JsonObject requestBody = new JsonObject();
-        requestBody.addProperty("userId", "randomUserId");
+        // using incorrect method
+        {
+            JsonObject requestBody = new JsonObject();
+            requestBody.addProperty("method", "invalidMethod");
+            requestBody.addProperty("token", "randomToken");
 
-        JsonObject response = io.supertokens.test.httpRequest.HttpRequest
-                .sendJsonPOSTRequest(process.getProcess(), "",
-                        "http://localhost:3567/recipe/user/password/reset/token", requestBody, 1000,
-                        1000,
-                        null, Utils.getCdiVersion2_6ForTests());
+            try {
+                io.supertokens.test.httpRequest.HttpRequest
+                        .sendJsonPOSTRequest(process.getProcess(), "",
+                                "http://localhost:3567/recipe/user/email/verify", requestBody, 1000,
+                                1000,
+                                null, Utils.getCdiVersion2_7ForTests());
+                throw new Exception("Should not have come here");
+            } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
+                assertTrue(e.statusCode == 400 && e.getMessage()
+                        .equals("Http error. Status Code: 400. Message: Unsupported method for email verification"));
+            }
+        }
 
-        assertEquals(response.get("status").getAsString(), "UNKNOWN_USER_ID_ERROR");
-        assertEquals(response.entrySet().size(), 1);
+        // passing invalid token
+        {
+            JsonObject requestBody = new JsonObject();
+            requestBody.addProperty("method", "token");
+            requestBody.addProperty("token", "invalidToken");
+
+            JsonObject response2 = io.supertokens.test.httpRequest.HttpRequest
+                    .sendJsonPOSTRequest(process.getProcess(), "",
+                            "http://localhost:3567/recipe/user/email/verify", requestBody, 1000,
+                            1000,
+                            null, Utils.getCdiVersion2_7ForTests());
+            assertEquals(response2.get("status").getAsString(), "EMAIL_VERIFICATION_INVALID_TOKEN_ERROR");
+            assertEquals(response2.entrySet().size(), 1);
+        }
 
         process.kill();
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
     }
+
 }
