+++ src/test/java/io/supertokens/test/VerifySessionAPITest2_7.java
@@ -28,7 +28,7 @@ import org.junit.rules.TestRule;
 import static junit.framework.TestCase.*;
 import static org.junit.Assert.assertNotNull;
 
-public class VerifySessionAPITest2_6 {
+public class VerifySessionAPITest2_7 {
     @Rule
     public TestRule watchman = Utils.getOnFailure();
 
@@ -64,7 +64,7 @@ public class VerifySessionAPITest2_6 {
         JsonObject sessionInfo = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "",
                         "http://localhost:3567/recipe/session", sessionRequest, 1000, 1000, null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
 
         JsonObject request = new JsonObject();
         request.addProperty("accessToken", sessionInfo.get("accessToken").getAsJsonObject().get("token").getAsString());
@@ -75,7 +75,7 @@ public class VerifySessionAPITest2_6 {
                 .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session/verify", request,
                         1000,
                         1000,
-                        null, Utils.getCdiVersion2_6ForTests());
+                        null, Utils.getCdiVersion2_7ForTests());
 
         assertEquals(response.get("status").getAsString(), "OK");
 
@@ -114,7 +114,7 @@ public class VerifySessionAPITest2_6 {
         JsonObject sessionInfo = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "",
                         "http://localhost:3567/recipe/session", sessionRequest, 1000, 1000, null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
 
         JsonObject refreshRequest = new JsonObject();
         refreshRequest.addProperty("refreshToken",
@@ -122,7 +122,7 @@ public class VerifySessionAPITest2_6 {
         refreshRequest.addProperty("enableAntiCsrf", false);
         sessionInfo = io.supertokens.test.httpRequest.HttpRequest.sendJsonPOSTRequest(process.getProcess(), "",
                 "http://localhost:3567/recipe/session/refresh", refreshRequest, 1000, 1000, null,
-                Utils.getCdiVersion2_6ForTests());
+                Utils.getCdiVersion2_7ForTests());
 
         JsonObject request = new JsonObject();
         request.addProperty("accessToken", sessionInfo.get("accessToken").getAsJsonObject().get("token").getAsString());
@@ -131,7 +131,7 @@ public class VerifySessionAPITest2_6 {
         JsonObject response = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session/verify", request,
                         1000,
-                        1000, null, Utils.getCdiVersion2_6ForTests());
+                        1000, null, Utils.getCdiVersion2_7ForTests());
 
         assertEquals(response.get("status").getAsString(), "OK");
 
@@ -177,7 +177,7 @@ public class VerifySessionAPITest2_6 {
         JsonObject sessionInfo = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "",
                         "http://localhost:3567/recipe/session", sessionRequest, 1000, 1000, null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
 
         JsonObject refreshRequest = new JsonObject();
         refreshRequest.addProperty("refreshToken",
@@ -185,7 +185,7 @@ public class VerifySessionAPITest2_6 {
         refreshRequest.addProperty("enableAntiCsrf", false);
         sessionInfo = io.supertokens.test.httpRequest.HttpRequest.sendJsonPOSTRequest(process.getProcess(), "",
                 "http://localhost:3567/recipe/session/refresh", refreshRequest, 1000, 1000, null,
-                Utils.getCdiVersion2_6ForTests());
+                Utils.getCdiVersion2_7ForTests());
 
         JsonObject request = new JsonObject();
         request.addProperty("accessToken", sessionInfo.get("accessToken").getAsJsonObject().get("token").getAsString());
@@ -194,7 +194,7 @@ public class VerifySessionAPITest2_6 {
         JsonObject response = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session/verify", request,
                         1000,
-                        1000, null, Utils.getCdiVersion2_6ForTests());
+                        1000, null, Utils.getCdiVersion2_7ForTests());
 
         assertEquals(response.get("status").getAsString(), "OK");
 
@@ -242,7 +242,7 @@ public class VerifySessionAPITest2_6 {
         JsonObject sessionInfo = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "",
                         "http://localhost:3567/recipe/session", sessionRequest, 1000, 1000, null,
-                        Utils.getCdiVersion2_6ForTests());
+                        Utils.getCdiVersion2_7ForTests());
 
         String sessionRemoveBodyString = "{" +
                 " sessionHandles : [ " + sessionInfo.get("session").getAsJsonObject().get("handle").getAsString() +
@@ -252,7 +252,7 @@ public class VerifySessionAPITest2_6 {
 
         io.supertokens.test.httpRequest.HttpRequest.sendJsonPOSTRequest(process.getProcess(), "",
                 "http://localhost:3567/recipe/session/remove", deleteRequest, 1000, 1000, null,
-                Utils.getCdiVersion2_6ForTests());
+                Utils.getCdiVersion2_7ForTests());
 
         JsonObject request = new JsonObject();
         request.addProperty("accessToken", sessionInfo.get("accessToken").getAsJsonObject().get("token").getAsString());
@@ -261,7 +261,7 @@ public class VerifySessionAPITest2_6 {
         JsonObject response = io.supertokens.test.httpRequest.HttpRequest
                 .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session/verify", request,
                         1000,
-                        1000, null, Utils.getCdiVersion2_6ForTests());
+                        1000, null, Utils.getCdiVersion2_7ForTests());
 
         assertEquals(response.get("status").getAsString(), "UNAUTHORISED");
         assertEquals(response.get("message").getAsString(), "Either the session has ended or has been blacklisted");
@@ -292,7 +292,7 @@ public class VerifySessionAPITest2_6 {
                 .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session/verify", request,
                         1000,
                         1000,
-                        null, Utils.getCdiVersion2_6ForTests());
+                        null, Utils.getCdiVersion2_7ForTests());
 
         assertEquals(response.get("status").getAsString(), "TRY_REFRESH_TOKEN");
         assertEquals(response.get("message").getAsString(), "io.supertokens.session.jwt.JWT$JWTException: Invalid JWT");
@@ -326,7 +326,7 @@ public class VerifySessionAPITest2_6 {
             JsonObject sessionInfo = io.supertokens.test.httpRequest.HttpRequest
                     .sendJsonPOSTRequest(process.getProcess(), "",
                             "http://localhost:3567/recipe/session", sessionRequest, 1000, 1000, null,
-                            Utils.getCdiVersion2_6ForTests());
+                            Utils.getCdiVersion2_7ForTests());
 
             JsonObject request = new JsonObject();
             request.addProperty("doAntiCsrfCheck", true);
@@ -334,7 +334,7 @@ public class VerifySessionAPITest2_6 {
             io.supertokens.test.httpRequest.HttpRequest
                     .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session/verify",
                             request,
-                            1000, 1000, null, Utils.getCdiVersion2_6ForTests());
+                            1000, 1000, null, Utils.getCdiVersion2_7ForTests());
             fail();
         } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
             assertEquals(e.statusCode, 400);
@@ -351,7 +351,7 @@ public class VerifySessionAPITest2_6 {
             JsonObject sessionInfo = io.supertokens.test.httpRequest.HttpRequest
                     .sendJsonPOSTRequest(process.getProcess(), "",
                             "http://localhost:3567/recipe/session", sessionRequest, 1000, 1000, null,
-                            Utils.getCdiVersion2_6ForTests());
+                            Utils.getCdiVersion2_7ForTests());
 
             JsonObject request = new JsonObject();
             request.addProperty("accessToken",
@@ -360,7 +360,7 @@ public class VerifySessionAPITest2_6 {
             io.supertokens.test.httpRequest.HttpRequest
                     .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session/verify",
                             request,
-                            1000, 1000, null, Utils.getCdiVersion2_6ForTests());
+                            1000, 1000, null, Utils.getCdiVersion2_7ForTests());
             fail();
         } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
             assertEquals(e.statusCode, 400);
@@ -377,7 +377,7 @@ public class VerifySessionAPITest2_6 {
             JsonObject sessionInfo = io.supertokens.test.httpRequest.HttpRequest
                     .sendJsonPOSTRequest(process.getProcess(), "",
                             "http://localhost:3567/recipe/session", sessionRequest, 1000, 1000, null,
-                            Utils.getCdiVersion2_6ForTests());
+                            Utils.getCdiVersion2_7ForTests());
 
             JsonObject request = new JsonObject();
             request.addProperty("accessToken",
@@ -387,7 +387,7 @@ public class VerifySessionAPITest2_6 {
             io.supertokens.test.httpRequest.HttpRequest
                     .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session/verify",
                             request,
-                            1000, 1000, null, Utils.getCdiVersion2_6ForTests());
+                            1000, 1000, null, Utils.getCdiVersion2_7ForTests());
             fail();
         } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
             assertEquals(e.statusCode, 400);
@@ -404,7 +404,7 @@ public class VerifySessionAPITest2_6 {
             JsonObject sessionInfo = io.supertokens.test.httpRequest.HttpRequest
                     .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session",
                             sessionRequest,
-                            1000, 1000, null, Utils.getCdiVersion2_6ForTests());
+                            1000, 1000, null, Utils.getCdiVersion2_7ForTests());
 
             JsonObject request = new JsonObject();
             request.addProperty("accessToken",
@@ -414,7 +414,7 @@ public class VerifySessionAPITest2_6 {
             JsonObject response = io.supertokens.test.httpRequest.HttpRequest
                     .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session/verify",
                             request,
-                            1000, 1000, null, Utils.getCdiVersion2_6ForTests());
+                            1000, 1000, null, Utils.getCdiVersion2_7ForTests());
 
             assertEquals(response.get("status").getAsString(), "TRY_REFRESH_TOKEN");
             assertEquals(response.get("message").getAsString(), "anti-csrf check failed");
@@ -427,7 +427,7 @@ public class VerifySessionAPITest2_6 {
             response = io.supertokens.test.httpRequest.HttpRequest
                     .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session/verify",
                             request,
-                            1000, 1000, null, Utils.getCdiVersion2_6ForTests());
+                            1000, 1000, null, Utils.getCdiVersion2_7ForTests());
             assertEquals(response.get("status").getAsString(), "OK");
         }
 
@@ -440,7 +440,7 @@ public class VerifySessionAPITest2_6 {
             JsonObject sessionInfo = io.supertokens.test.httpRequest.HttpRequest
                     .sendJsonPOSTRequest(process.getProcess(), "",
                             "http://localhost:3567/recipe/session", sessionRequest, 1000, 1000, null,
-                            Utils.getCdiVersion2_6ForTests());
+                            Utils.getCdiVersion2_7ForTests());
 
             JsonObject request = new JsonObject();
             request.addProperty("accessToken",
@@ -451,7 +451,7 @@ public class VerifySessionAPITest2_6 {
                     .HttpRequest
                     .sendJsonPOSTRequest(process.getProcess(), "", "http://localhost:3567/recipe/session/verify",
                             request,
-                            1000, 1000, null, Utils.getCdiVersion2_6ForTests());
+                            1000, 1000, null, Utils.getCdiVersion2_7ForTests());
             fail();
         } catch (io.supertokens.test.httpRequest.HttpResponseException e) {
             assertEquals(e.statusCode, 400);
